LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select-multiple-popup/select-multiple-popup-keyboard-behavior.tentative-expected.txt



FAIL Keyboard behavior for base appearance <select multiple size=1> assert_equals: appearance:base-select must be supported to run this test. expected "base-select" but got "auto"
